refactor and hero
