+++ src/components/Animated-steps-list/hero/index.js
@@ -1,5 +1,4 @@
 import React from "react";
-import {Link} from "gatsby";
 import Button from "../../../reusecore/Button";
 import { HeroWrapper } from "./hero.style";
 
@@ -8,15 +7,13 @@ const Hero = ({ title, description, subtitle, links }) => {
     return(
         <HeroWrapper>
             <div className="content">
+                <h5 className="subtitle">{subtitle}</h5>
                 <h1>{title}</h1>
-                <p className="subtitle">{subtitle}</p>
-                <p className="description">{description}</p>
+                <p>{description}</p>
                 <div className="links">
                     {links.map((link) => {
                         return (
-                            <Link key={link.text} to={link.url}>
-                                <Button secondary>{link.text}</Button>
-                            </Link>
+                            <Button secondary key={link.text} title={link.text} url={link.url}/>
                         );
                     })}
                 </div>
